// src/enums/swapiUrls.ts
export enum swapiUrls {
    CHARACTERS = 'https://swapi.dev/api/people',
    MOVIES = 'https://swapi.dev/api/films',
    PLANETS = 'https://swapi.dev/api/planets',
    SPECIES = 'https://swapi.dev/api/species',
    STARSHIPS = 'https://swapi.dev/api/starships',
    VEHICLES = 'https://swapi.dev/api/vehicles'
};